move array parameter to its own module
